Canonicalize given paths to allow non-absolute paths as args.
Currently, the tool chdir()'s but also uses paths that are created
simply by appending the given command-line arguments with suffixes. This
works fine if the user provides absolute paths on the command line, but
breaks otherwise with file-not-found errors. This patch uses
`std::fs::canonicalize()` to normalize all given paths first.